refactor: move comparison release functions to transformations.py
