+++ src/core/transformations.py
@@ -3,7 +3,10 @@ from enum import IntEnum
 import numpy as np
 
 from util.check import Checker
-from util.exceptions import InvalidGainInterpretationValue
+from util.exceptions import (
+    InvalidGainInterpretationValue,
+    ReleasePlannedAndDevelopedOfDifferentSizes,
+)
 
 
 class GainInterpretation(IntEnum):
@@ -62,3 +65,27 @@ def calculate_aggregated_weighted_value(values, weights):
     ) / np.linalg.norm(weights)
 
     return aggregated_weighted_value
+
+
+def normDiff(rp, rd):
+    """The normDiff function performs a vector transformation between the planned release vector and the redeveloped
+    release vector. This transformation represents the quantitative perception of the difference between the
+    planned quality requirement for a release, and the observed result, after its development.
+    """
+
+    if len(rp) != len(rd):
+        raise ReleasePlannedAndDevelopedOfDifferentSizes(
+            "The size between planned and developed release vectors is not equal.",
+        )
+    return np.linalg.norm(rp - rd) / np.linalg.norm(rp)
+
+
+def diff(rp, rd):
+    """The diff function interprets a vector transformation between the planned and developed release vectors.
+    It generates a vector that expresses whether the result observed in a release is above or below the planned quality
+    requirement."""
+    if len(rp) != len(rd):
+        raise ReleasePlannedAndDevelopedOfDifferentSizes(
+            "The size between planned and developed release vectors is not equal.",
+        )
+    return [max(0, x - y) for x, y in zip(rp, rd)]
