+++ src/util/exceptions.py
@@ -54,6 +54,7 @@ class MeasureKeyNotSupported(MeasureSoftGramCoreException):
 
     pass
 
+
 class ReleasePlannedAndDevelopedOfDifferentSizes(MeasureSoftGramCoreException):
     """Raised when the sizes of the planned and developed release vectors are different"""
